[WIP] 536 - running stream operators
