+++ otel-extension/src/main/scala/io/scalac/mesmer/otelextension/instrumentations/akka/stream/AkkaStreamMonitorExtension.scala
@@ -16,12 +16,15 @@ import scala.concurrent.duration.DurationInt
 import scala.util.Failure
 import scala.util.Success
 
+import io.scalac.mesmer.core.akka.stream.ConnectionsIndexCache
 import io.scalac.mesmer.core.cluster.ClusterNode.ActorSystemOps
 import io.scalac.mesmer.core.event.Service.streamService
 import io.scalac.mesmer.core.event.StreamEvent
 import io.scalac.mesmer.core.event.StreamEvent.LastStreamStats
 import io.scalac.mesmer.core.event.StreamEvent.StreamInterpreterStats
 import io.scalac.mesmer.core.model.ShellInfo
+import io.scalac.mesmer.core.model.stream.ConnectionStats
+import io.scalac.mesmer.core.model.stream.StageInfo
 import io.scalac.mesmer.core.util.Retry
 import io.scalac.mesmer.core.util.stream
 import io.scalac.mesmer.otelextension.instrumentations.akka.stream.AkkaStreamMonitorExtension.StreamStatsReceived
@@ -32,6 +35,8 @@ final class AkkaStreamMonitorExtension(actorSystem: ActorSystem[_]) extends Exte
   private val metrics = new AkkaStreamMetrics(node)
 
   private val interval = AkkaStreamConfig.metricSnapshotRefreshInterval(actorSystem.classicSystem)
+  // FIXME: move to config
+  private val indexCache = ConnectionsIndexCache.bounded(100)
 
   def start(): Behavior[StreamStatsReceived] = Behaviors.setup[StreamStatsReceived] { ctx =>
     actorSystem.receptionist ! Register(
@@ -64,7 +69,38 @@ final class AkkaStreamMonitorExtension(actorSystem: ActorSystem[_]) extends Exte
     val streamNames = streamShells.keySet.map(_.streamName)
     metrics.setRunningStreamsTotal(streamNames.size)
     metrics.setRunningActorsTotal(current.size)
+
+    /*
+     FIXME:
+     - process "current" shell infos based on how it's done in `AkkaStreamMonitoring`
+     - ConnectionsIndexCache is now in `core`
+     - based on the processed shells, create `metrics.runningOperators` metric
+     - compare reported metrics to the version 0.7.0
+     - if it doesn't match, try to append shellInfos instead of replacing them in lines 50 and 53
+     */
+
   }
+
+  private def computeSnapshotEntries(
+    stage: StageInfo,
+    connectionStats: Set[ConnectionStats],
+    distinct: Boolean,
+    extractFunction: ConnectionStats => Int
+  ): Seq[StageInfo] =
+    // FIXME: Simplify
+    if (connectionStats.nonEmpty) {
+      if (distinct) {
+        // optimization for simpler graphs
+        connectionStats.toSeq.map(_ => stage)
+      } else {
+        connectionStats
+          .map(extractFunction)
+          .toSeq
+          .map(_ => stage)
+      }
+    } else {
+      Seq(stage)
+    }
 
   actorSystem.systemActorOf(start(), "mesmerStreamMonitor")
